Move test classes to the correct packages
