+++ src/test/java/me/zeroeightsix/fiber/impl/tree/ConfigQueryTest.java
@@ -1,4 +1,4 @@
-package me.zeroeightsix.fiber.tree;
+package me.zeroeightsix.fiber.impl.tree;
 
 import me.zeroeightsix.fiber.api.exception.FiberQueryException;
 import me.zeroeightsix.fiber.api.tree.ConfigLeaf;
